Inline typescript private configs into public typescript.js
Consolidate typescript-airbnb, typescript-rules, and typescript-utils
into a single configs/public/typescript.js for clarity.
